fix(domain): eliminate domain→application layer import violations
- ports.rs: change test imports from application::http_client to
  domain::http_error/domain::http_port (where types are actually defined)
- entities.rs: move UrlDeduplicator test to application/deduplicator.rs
  (domain should not depend on application layer, even in tests)
